+++ google-finance/src/config.ts
@@ -1,4 +1,0 @@
-import { Requester } from '@chainlink/external-adapter'
-import { Config } from '@chainlink/types'
-
-export const makeConfig = (prefix?: string): Config => Requester.getDefaultConfig(prefix)