+++ google-finance/typings/boxhock_google-finance-data/index.d.ts
@@ -1,2 +1,0 @@
-// Declare missing type definitions
-declare module 'boxhock_google-finance-data'