+++ CHANGELOG.md
@@ -55,7 +55,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - `cryptoapis`
 - dxFeed now uses OTC feeds for FTSE and N225, rather than licensed data feeds
 - Migrated to Typescript
-  - `google-finance`
   - `kaiko`
   - `marketstack`
   - `metalsapi`
@@ -64,6 +63,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - `polygon`
   - `nomics`
   - `openexchangerates`
+- Removed
+  - `google-finance`
 
 ### Removed